Ahora se calcula el total de dinero generado por empleado con sus ventas

+++ src/main/java/ar/edu/unlam/tallerweb1/domain/informe/ServicioInforme.java
@@ -3,5 +3,5 @@ package ar.edu.unlam.tallerweb1.domain.informe;
 import java.util.Map;
 
 public interface ServicioInforme {
-    Map<Long, Integer> obtenerVentasPorEmpleado();
+    Map<Long, Double> obtenerVentasPorEmpleado();
 }
